requirements.txt
mediapipe==0.10.0
opencv-python>4.7.0
dbus-python==1.3.2
